+++ src/kaleidoscope/device/softhruf/Splitography.cpp
@@ -1,13 +1,12 @@
 /* -*- mode: c++ -*-
- * Kaleidoscope-Hardware-EZ-ErgoDox -- ErgoDox hardware support for Kaleidoscope
- * Copyright (C) 2018  Keyboard.io, Inc
+ * Kaleidoscope-Hardware-SOFTHRUF-Splitography -- Splitography hardware support for Kaleidoscope
+ * Copyright (C) 2018, 2019  Keyboard.io, Inc
  *
- * Based on QMK (commit 8cdb4a915)
- *  (C) Jack Humbert, Erez Zukerman, Oleg Kostyuk
+ * Based on QMK (commit e9a67f8fd) and sdothum's fork (commit 8616b44)
+ *  (C) Jack Humbert, Jun Wako, Steven Hum, and others
  * Original QMK sources:
- *  - keyboards/ergodox_ez/ergodox_ez.c
- *  - keyboards/ergodox_ez/ergodox_ez.h
- *  - keyboards/ergodox_ez/matrix.c
+ *  - keyboards/splitography/config.h
+ *  - keyboards/splitography/splitography.h
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -23,35 +22,23 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#pragma once
+#ifdef ARDUINO_AVR_SPLITOGRAPHY
 
-#ifdef ARDUINO_AVR_ERGODOX
-
-#include <Arduino.h>
+#include <Kaleidoscope.h>
+#include <avr/wdt.h>
 
 namespace kaleidoscope {
-namespace hardware {
-namespace ez {
+namespace device {
+namespace softhruf {
 
-class ErgoDoxScanner {
- public:
-  ErgoDoxScanner() {}
-
-  void begin();
-  void toggleATMegaRow(int row);
-  void selectExtenderRow(int row);
-  uint8_t readCols(int row);
-
-  void reattachExpanderOnError();
-
- private:
-  uint8_t expander_error_ = 0x20;
-
-  uint8_t initExpander();
-};
+ATMEGA_KEYBOARD_DATA(Splitography);
+constexpr int8_t Splitography::led_count;
 
 }
 }
 }
+
+HARDWARE_IMPLEMENTATION KeyboardHardware;
+kaleidoscope::device::softhruf::Splitography &Splitography = KeyboardHardware;
 
 #endif
