+++ src/kaleidoscope/device/technomancy/Atreus2.h
@@ -21,17 +21,17 @@
 #ifdef ARDUINO_AVR_ATREUS2
 
 #include <Arduino.h>
-#define HARDWARE_IMPLEMENTATION kaleidoscope::hardware::technomancy::Atreus2
+#define HARDWARE_IMPLEMENTATION kaleidoscope::device::technomancy::Atreus2
 #include "Kaleidoscope-HIDAdaptor-KeyboardioHID.h"
 
 #include "kaleidoscope/macro_helpers.h"
 
-#include "kaleidoscope/hardware/ATMegaKeyboard.h"
+#include "kaleidoscope/device/ATMegaKeyboard.h"
 
 namespace kaleidoscope {
-namespace hardware {
+namespace device {
 namespace technomancy {
-class Atreus2: public kaleidoscope::hardware::ATMegaKeyboard {
+class Atreus2: public kaleidoscope::device::ATMegaKeyboard {
   friend class ATMegaKeyboard;
  public:
   Atreus2(void) {}
@@ -82,8 +82,8 @@ class Atreus2: public kaleidoscope::hardware::ATMegaKeyboard {
 }
 }
 
-#include "kaleidoscope/hardware/key_indexes.h"
+#include "kaleidoscope/device/key_indexes.h"
 
-extern kaleidoscope::hardware::technomancy::Atreus2 &Atreus2;
+extern kaleidoscope::device::technomancy::Atreus2 &Atreus2;
 
 #endif
